+++ backend/controllers/ComentariosutilizadorController.php
@@ -8,18 +8,37 @@ use app\models\ComentariosutilizadorSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
 
 /**
  * ComentariosutilizadorController implements the CRUD actions for Comentariosutilizador model.
  */
-class ComentariosutilizadorController extends Controller
-{
+class ComentariosutilizadorController extends Controller {
+
     /**
      * {@inheritdoc}
      */
-    public function behaviors()
-    {
+    public function behaviors() {
         return [
+            'access' => [
+                'class' => AccessControl::classname(),
+                'only' => ['create', 'update', 'delete', 'login', 'logout'],
+                'rules' => [
+                        [
+                        'allow' => true,
+                        'actions' => ['logout', 'create', 'update', 'delete', 'login'],
+                        'roles' => ['@']
+                    ],
+                        [
+                        'allow' => true,
+                        'actions' => ['login'],
+                        'roles' => ['?']
+                    ],
+                    'denyCallback' => function ($rule, $action) {
+                        throw new \Exception('Não tem permissão para aceder a esta página.');
+                    }
+                ]
+            ],
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -33,15 +52,18 @@ class ComentariosutilizadorController extends Controller
      * Lists all Comentariosutilizador models.
      * @return mixed
      */
-    public function actionIndex()
-    {
-        $searchModel = new ComentariosutilizadorSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+    public function actionIndex() {
+        if (Yii::$app->user->can('admin') or Yii::$app->user->can('admin')) {
+            $searchModel = new ComentariosutilizadorSearch();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                        'searchModel' => $searchModel,
+                        'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
@@ -51,11 +73,14 @@ class ComentariosutilizadorController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
-    public function actionView($Id_comentario, $Id_utilizador)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($Id_comentario, $Id_utilizador),
-        ]);
+    public function actionView($Id_comentario, $Id_utilizador) {
+        if (Yii::$app->user->can('admin') or Yii::$app->user->can('admin')) {
+            return $this->render('view', [
+                        'model' => $this->findModel($Id_comentario, $Id_utilizador),
+            ]);
+        } else {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
@@ -63,17 +88,20 @@ class ComentariosutilizadorController extends Controller
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
-    public function actionCreate()
-    {
-        $model = new Comentariosutilizador();
+    public function actionCreate() {
+        if (Yii::$app->user->can('admin') or Yii::$app->user->can('admin')) {
+            $model = new Comentariosutilizador();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'Id_comentario' => $model->Id_comentario, 'Id_utilizador' => $model->Id_utilizador]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'Id_comentario' => $model->Id_comentario, 'Id_utilizador' => $model->Id_utilizador]);
+            }
+
+            return $this->render('create', [
+                        'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException;
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -84,17 +112,20 @@ class ComentariosutilizadorController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
-    public function actionUpdate($Id_comentario, $Id_utilizador)
-    {
-        $model = $this->findModel($Id_comentario, $Id_utilizador);
+    public function actionUpdate($Id_comentario, $Id_utilizador) {
+        if (Yii::$app->user->can('admin') or Yii::$app->user->can('admin')) {
+            $model = $this->findModel($Id_comentario, $Id_utilizador);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'Id_comentario' => $model->Id_comentario, 'Id_utilizador' => $model->Id_utilizador]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'Id_comentario' => $model->Id_comentario, 'Id_utilizador' => $model->Id_utilizador]);
+            }
+
+            return $this->render('update', [
+                        'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException;
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -105,11 +136,14 @@ class ComentariosutilizadorController extends Controller
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
      */
-    public function actionDelete($Id_comentario, $Id_utilizador)
-    {
-        $this->findModel($Id_comentario, $Id_utilizador)->delete();
+    public function actionDelete($Id_comentario, $Id_utilizador) {
+        if (Yii::$app->user->can('admin') or Yii::$app->user->can('admin')) {
+            $this->findModel($Id_comentario, $Id_utilizador)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new ForbiddenHttpException;
+        }
     }
 
     /**
@@ -120,12 +154,12 @@ class ComentariosutilizadorController extends Controller
      * @return Comentariosutilizador the loaded model
      * @throws NotFoundHttpException if the model cannot be found
      */
-    protected function findModel($Id_comentario, $Id_utilizador)
-    {
+    protected function findModel($Id_comentario, $Id_utilizador) {
         if (($model = Comentariosutilizador::findOne(['Id_comentario' => $Id_comentario, 'Id_utilizador' => $Id_utilizador])) !== null) {
             return $model;
         }
 
         throw new NotFoundHttpException('The requested page does not exist.');
     }
+
 }
